damo_reclaim/set_param(): skip writing if the parameter is not existing
'skip_anon' parameter has added from Linux v6.3.  On kernels older than
that, set_param() will crash.  Check existence of the file and skip.  To
clarify the situation, warn the finding.

+++ src/damo_reclaim.py
@@ -38,6 +38,9 @@ def set_param(param, val):
     if val == None:
         return
     path = os.path.join(darc_params_dir, param)
+    if not os.path.isfile(path):
+        print('warn: %s not exist; setup of it is skipped' % path)
+        return
     with open(path, 'w') as f:
         f.write('%s' % val)
